Clean up repo Remove tset files, older calendar update runme.py add official calendar
