+++ .gitignore
@@ -2,3 +2,4 @@
 poll_results.json
 scratchpad.txt
 poetry.lock
+simple.py
