+++ Formula_1.ics
@@ -9,7 +9,7 @@ LAST-MODIFIED:20230411T210909Z
 DTSTAMP:20230411T210909Z
 LOCATION:Canada
 X-ECAL-SCHEDULE:63ed7ad04c410f0c65362129
-DTSTART:20230417T082000Z
+DTSTART:20230617T200000Z
 DTEND:20230617T210000Z
 SUMMARY:⏱️ FORMULA 1 PIRELLI GRAND PRIX DU CANADA 2023 - Qualifying
 TRANSP:TRANSPARENT
@@ -2055,7 +2055,7 @@ LAST-MODIFIED:20230411T210909Z
 DTSTAMP:20230411T210909Z
 LOCATION:Canada
 X-ECAL-SCHEDULE:63ed7ad04c410f0c65362129
-DTSTART:20230416T052000Z
+DTSTART:20230616T173000Z
 DTEND:20230616T183000Z
 SUMMARY:🏎 FORMULA 1 PIRELLI GRAND PRIX DU CANADA 2023 - Practice 1
 TRANSP:TRANSPARENT
